@@ -12,42 +12,39 @@
     "generate": "nuxt generate",
     "preview": "npx nuxi preview",
     "postinstall": "nuxt prepare",
-    "lint:js": "eslint --ext \".ts,.vue\" --ignore-path .eslintignore .",
-    "lint:prettier": "prettier --check --write .",
-    "lint:fix": "eslint --ext \".ts,.vue\" --ignore-path .eslintignore . --fix"
+    "lint:js": "npx eslint --ignore-pattern \"recruitis-shared/*\" .",
+    "lint:fix": "npx eslint . --fix"
   },
   "devDependencies": {
     "@css-render/vue3-ssr": "0.15.14",
-    "@iconify-json/ion": "^1.1.18",
+    "@eslint-community/eslint-plugin-eslint-comments": "^4.4.0",
+    "@eslint/js": "^9.9.0",
+    "@iconify-json/ion": "1.1.19",
     "@iconify/vue": "^4.1.2",
-    "@kyvg/vue3-notification": "^2.9.1",
-    "@nuxt/devtools": "1.3.9",
-    "@nuxt/icon": "^1.4.0",
-    "@nuxtjs/eslint-config": "^12.0.0",
-    "@nuxtjs/eslint-config-typescript": "12.1.0",
-    "@nuxtjs/eslint-module": "^4.1.0",
+    "@kyvg/vue3-notification": "3.2.1",
+    "@nuxt/devtools": "1.3.14",
+    "@nuxt/eslint": "^0.5.1",
+    "@nuxt/icon": "1.4.7",
     "@nuxtjs/google-fonts": "3.2.0",
-    "@nuxtjs/i18n": "8.3.3",
+    "@nuxtjs/i18n": "8.5.0",
     "@nuxtjs/strapi": "1.12.0",
     "@primevue/nuxt-module": "^4.0.4",
-    "@sidebase/nuxt-auth": "0.8.1",
+    "@sidebase/nuxt-auth": "0.8.2",
     "@tailwindcss/container-queries": "^0.1.1",
     "@tailwindcss/forms": "0.5.7",
-    "@tailwindcss/typography": "0.5.13",
+    "@tailwindcss/typography": "0.5.14",
+    "@types/eslint__js": "^8.42.3",
     "@types/node": "22.0.0",
     "@typescript-eslint/eslint-plugin": "7.18.0",
     "@typescript-eslint/parser": "7.18.0",
     "@vueform/multiselect": "2.6.9",
     "@vuepic/vue-datepicker": "6.1.0",
-    "@vueuse/core": "10.11.0",
-    "autoprefixer": "10.4.19",
-    "date-fns": "2.30.0",
-    "eslint": "^8.48.0",
+    "@vueuse/core": "11.0.1",
+    "autoprefixer": "10.4.20",
+    "date-fns": "3.6.0",
+    "eslint": "^8.57.0",
     "eslint-config-airbnb-typescript": "18.0.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-eslint-comments": "^3.2.0",
     "eslint-plugin-nuxt": "^4.0.0",
-    "eslint-plugin-prettier": "5.2.1",
     "eslint-plugin-storybook": "0.8.0",
     "eslint-plugin-tailwindcss": "^3.17.4",
     "eslint-plugin-vue": "9.27.0",
