@@ -1,13 +1,15 @@
 <script setup lang="ts">
 useHead({
     title: 'Nuxt starter',
-});
+})
 </script>
+
 <template>
     <NuxtLayout>
         <NuxtPage></NuxtPage>
     </NuxtLayout>
 </template>
+
 <style lang="pcss">
 body {
     @apply font-sans text-gray-900;
