@@ -1,7 +1,7 @@
-import { defineNuxtRouteMiddleware } from '#app';
+import { defineNuxtRouteMiddleware } from '#app'
 
 export default defineNuxtRouteMiddleware(async (to) => {
     if (to.path.startsWith('/protected/')) {
-        await useAuth();
+        await useAuth()
     }
-});
+})
