@@ -1,27 +1,33 @@
 <script lang="ts" setup>
-import Page from '~/sites/default/pages/page.vue';
-const { locale } = useI18n();
-const route = useRoute();
-const slug = Array.isArray(route.params.slug) ? route.params.slug[0] : route.params.slug || 'index';
+import Page from '~/sites/default/pages/page.vue'
 
-const { data: pageData } = await useAsyncData(() =>
-    $fetch('/api/page', {
-        params: { locale: locale.value, slug, homepage: false },
-    }),
-);
+const { locale } = useI18n()
+const route = useRoute()
+const slug = Array.isArray(route.params.slug) ? route.params.slug[0] : route.params.slug || 'index'
 
-const page = pageData.value;
+const { data: pageData } = await useAsyncData(() => $fetch(
+    '/api/page',
+    {
+        params: { locale: locale.value,
+            slug,
+            homepage: false },
+    },
+))
+
+const page = pageData.value
 
 useHead({
     title: page?.title,
-});
+})
 </script>
+
 <template>
     <Page
         v-if="page"
         v-bind="page"
     ></Page>
-    <template v-else>
-        <div>template 404 - neexistuje</div>
-    </template>
+
+    <div v-else>
+        template 404 - neexistuje
+    </div>
 </template>
