@@ -1,10 +1,12 @@
 <script setup lang="ts">
 </script>
+
 <template>
     <div class="inline-flex items-center justify-center rounded-full bg-white p-2.5 shadow-md shadow-black/10 ring-1 ring-black/5">
-        <div class="loader h-4 w-4 border-2 border-[var(--tw-brand-500)]"></div>
+        <div class="loader size-4 border-2 border-[var(--tw-brand-500)]"></div>
     </div>
 </template>
+
 <style scoped>
 .loader {
   aspect-ratio: 1;
