@@ -1,3 +1,3 @@
-export { default as MultiSelect } from './MultiSelect.vue';
-export { MultiselectStyles } from './styles';
-export type { IMultiselectProps, IOption } from './types';
+export { default as MultiSelect } from './MultiSelect.vue'
+export { MultiselectStyles } from './styles'
+export type { IMultiselectProps, IOption } from './types'
