@@ -1,16 +1,17 @@
-import { type MultiSelectProps } from 'primevue/multiselect';
+import type { MultiSelectProps } from 'primevue/multiselect'
+
 export interface IOption {
-    label: string;
-    value: string;
+    label: string
+    value: string
 }
 
 export interface IMultiselectProps extends MultiSelectProps {
-    id: string;
-    name?: string;
-    initialValue?: IOption[] | null | undefined;
-    options: IOption[];
-    closeOnSelect?: boolean;
+    id: string
+    name?: string
+    initialValue?: IOption[] | null | undefined
+    options: IOption[]
+    closeOnSelect?: boolean
     lang?: {
-        placeholder?: string;
+        placeholder?: string
     }
 }
