@@ -1,9 +1,14 @@
 <script lang="ts" setup>
-import Select from '~/recruitis-shared/components/forms/select.vue';
-import type { SelectInterface } from '~/recruitis-shared/components/forms/select.vue';
-const emit = defineEmits(['update:initialValue']);
-const props = withDefaults(defineProps<SelectInterface>(), {});
+import Select from '~/recruitis-shared/components/forms/select.vue'
+import type { SelectInterface } from '~/recruitis-shared/components/forms/select.vue'
+
+const emit = defineEmits(['update:initialValue'])
+const props = withDefaults(
+    defineProps<SelectInterface>(),
+    {},
+)
 </script>
+
 <template>
     <Select
         v-bind="props"
