@@ -1,47 +1,51 @@
 <script setup lang="ts">
-import { JobsFilter, Jobs } from '~/sites/default/components/blocks/jobsList';
-import type { IJobsListBlock } from '~/sites/default/types/pages';
-import type { Types } from '~/sites/default/components/blocks/jobsList';
-import { useJobsData } from '~/composables/useJobsData';
-const props = defineProps<IJobsListBlock>();
-const { data: jobs, refresh, status } = useJobsData<{ data: Types.IJob[]; meta: { entries_total: number }  }>();
+import type { Types, JobsFilter, type Jobs } from '~/sites/default/components/blocks/jobsList'
+import type { IJobsListBlock } from '~/sites/default/types/pages'
+import { useJobsData } from '~/composables/useJobsData'
 
-const route = useRoute();
+const props = defineProps<IJobsListBlock>()
+const { data: jobs, refresh, status } = useJobsData<{ data: Types.IJob[]
+    meta: { entries_total: number } }>()
+
+const route = useRoute()
 
 const scrollToAnchor = (limit: string, smooth: boolean = true) => {
-    const element = document.getElementById(`anchor-${limit}`);
+    const element = document.getElementById(`anchor-${limit}`)
     if (element) {
-        element.scrollIntoView({ behavior: smooth ? 'smooth' : 'auto' });
+        element.scrollIntoView({ behavior: smooth ? 'smooth' : 'auto' })
     }
-};
+}
 
 const onAfterRefresh = async (limit: number) => {
-    await refresh();
-    scrollToAnchor(limit.toString());
-};
+    await refresh()
+    scrollToAnchor(limit.toString())
+}
 
 // Scroll to anchor on page load if limit is present in query string
 onMounted(() => {
-    const limit = route.query.limit as string;
+    const limit = route.query.limit as string
     if (limit) {
-        scrollToAnchor(limit, false);
+        scrollToAnchor(
+            limit,
+            false,
+        )
     }
-});
-const filter = ref();
+})
+const filter = ref()
 watch(
     () => route.query,
     async (newQuery, oldQuery) => {
         if (Object.keys(newQuery).length === 0 && JSON.stringify(newQuery) !== JSON.stringify(oldQuery)) {
             if (filter.value) {
-                await filter.value?.clearFormState();
+                await filter.value?.clearFormState()
             }
-            await refresh();
+            await refresh()
         }
     },
     { deep: true },
-);
-
+)
 </script>
+
 <template>
     <div class="container space-y-4">
         <div class="prose">
