@@ -1,4 +1,4 @@
-export { default as Button } from '~/sites/default/components/Button.vue';
-export { default as Checkboxes } from '~/sites/default/components/Checkboxes.vue';
-export { default as Input } from '~/sites/default/components/Input.vue';
-export { MultiSelect, MultiselectStyles } from '~/sites/default/components/Multiselect';
+export { default as Button } from '~/sites/default/components/Button.vue'
+export { default as Checkboxes } from '~/sites/default/components/Checkboxes.vue'
+export { default as Input } from '~/sites/default/components/Input.vue'
+export { MultiSelect, MultiselectStyles } from '~/sites/default/components/Multiselect'
