@@ -11,12 +11,17 @@ export const i18n = {
         },
     ],
     vueI18n: './sites/default/i18n.config.ts',
-};
+}
 export const googleFonts = {
     families: {
         Inter: {
-            wght: [400, 500, 600, 800],
+            wght: [
+                400,
+                500,
+                600,
+                800,
+            ],
         },
     },
-};
-export const componentsPaths = ['~/sites/default/components'];
+}
+export const componentsPaths = ['~/sites/default/components']
