@@ -1,32 +1,32 @@
 interface IMenuItemPageRelation {
     data: {
-        id: number;
+        id: number
         attributes: {
-            Title: string;
-            Url: string;
-            Homepage: boolean;
-        };
-    };
+            Title: string
+            Url: string
+            Homepage: boolean
+        }
+    }
 }
 export interface IMenuItem {
     attributes: {
-        title: string;
-        url: string;
-        target: string;
-        link_hidden: boolean;
+        title: string
+        url: string
+        target: string
+        link_hidden: boolean
         children?: {
             data: [
                 {
                     attributes: {
-                        title: string;
-                        url: string;
-                        target: string;
-                        link_hidden: boolean;
-                        page_relation: IMenuItemPageRelation;
-                    };
+                        title: string
+                        url: string
+                        target: string
+                        link_hidden: boolean
+                        page_relation: IMenuItemPageRelation
+                    }
                 },
-            ];
-        };
-        page_relation: IMenuItemPageRelation;
-    };
+            ]
+        }
+        page_relation: IMenuItemPageRelation
+    }
 }
