@@ -1,46 +1,46 @@
 interface IStrapiPageLocalizations {
-    data: IStrapiPage[];
+    data: IStrapiPage[]
 }
 
 export enum IStrapiBlockName {
     JobsList = 'blocks.jobs-list',
     AnotherComponent = 'blocks.another-component',
 }
-export type IStrapiBlockNameUnion = IStrapiBlockName.JobsList | IStrapiBlockName.AnotherComponent;
+export type IStrapiBlockNameUnion = IStrapiBlockName.JobsList | IStrapiBlockName.AnotherComponent
 
 export interface IStrapiBlock {
-    id: number;
-    __component: IStrapiBlockName;
+    id: number
+    __component: IStrapiBlockName
 }
 
 export interface IJobsListBlock extends IStrapiBlock {
-    showFilter?: boolean;
-    showSubmitButton?: boolean;
+    showFilter?: boolean
+    showSubmitButton?: boolean
 }
 
 export interface IStrapiAnotherComponentBlock extends IStrapiBlock {
-    showFilter?: boolean;
+    showFilter?: boolean
 }
 
-export type IStrapiBlockUnion = IJobsListBlock | IStrapiAnotherComponentBlock;
+export type IStrapiBlockUnion = IJobsListBlock | IStrapiAnotherComponentBlock
 
 export interface IStrapiPage {
     attributes: {
-        Homepage: boolean;
-        Description?: string | null;
-        Title: string;
-        locale: string;
-        localizations: IStrapiPageLocalizations;
-        Blocks: IStrapiBlockUnion[];
-        Url: string;
-        createdAt: string;
-        updatedAt: string;
-        publishedAt: string;
-    };
+        Homepage: boolean
+        Description?: string | null
+        Title: string
+        locale: string
+        localizations: IStrapiPageLocalizations
+        Blocks: IStrapiBlockUnion[]
+        Url: string
+        createdAt: string
+        updatedAt: string
+        publishedAt: string
+    }
 }
 
 export interface IPage {
-    title: string;
-    description?: string | null;
-    blocks?: IStrapiBlockUnion[];
+    title: string
+    description?: string | null
+    blocks?: IStrapiBlockUnion[]
 }
