@@ -7,15 +7,22 @@ export const i18n = {
         },
     ],
     vueI18n: './sites/default/i18n.config.ts',
-};
+}
 export const googleFonts = {
     families: {
         'Nunito+Sans': {
-            wght: [600, 700, 800],
+            wght: [
+                600,
+                700,
+                800,
+            ],
         },
         'DM+Serif+Display': {
             wght: [400],
         },
     },
-};
-export const componentsPaths = ['~/sites/default/components', '~/components'];
+}
+export const componentsPaths = [
+    '~/sites/default/components',
+    '~/components',
+]
